refactor: delete `miden-tx-kernel-sys` crate and move the code to `miden-base-sys`
Add `bindings` and `masm-lib` features to the `miden-base-sys`
crate to separate the the user facing Rust bindings from the MASM library
needed only for the compiler.
