+++ Cargo.toml
@@ -95,7 +95,6 @@ midenc-compile = { version = "0.0.1", path = "midenc-compile" }
 midenc-driver = { version = "0.0.1", path = "midenc-driver" }
 midenc-runner = { version = "0.0.1", path = "midenc-runner" }
 midenc-session = { version = "0.0.1", path = "midenc-session" }
-miden-base-sys = { version = "0.0.0", path = "sdk/base-sys" }   
 miden-integration-tests = { version = "0.0.0", path = "tests/integration" }
 wat = "1.0.69"
 blake3 = "1.5"
