--- sdk/Cargo.toml
@@ -1,27 +1,0 @@
-[workspace]
-resolver = "2"
-members = [
-  "stdlib-sys",
-  "tx-kernel-sys",
-  "sdk",
-]
-
-[workspace.package]
-
-version = "0.0.0"
-rust-version = "1.71"
-authors = ["Miden contributors"]
-repository = "https://github.com/0xPolygonMiden/compiler"
-homepage = "https://github.com/0xPolygonMiden/compiler"
-documentation = "https://github.com/0xPolygonMiden/compiler"
-categories = ["Compilers"]
-keywords = ["compiler", "miden"]
-license = "MIT"
-readme = "README.md"
-edition = "2021"
-
-
-[profile.release]
-panic = "abort"
-# optimize for size
-opt-level = "z"
