+++ sdk/base-sys/src/bindings/tx/mod.rs
@@ -1,5 +1,3 @@
-#![no_std]
-
 mod externs;
 use externs::*;
