+++ sdk/base-sys/src/lib.rs
@@ -1,24 +1,7 @@
-use miden_assembly::{library::Library as CompiledLibrary, utils::Deserializable};
+// Enable no_std for the bindings module
+#![cfg_attr(feature = "bindings", no_std)]
 
-/// Stubs for the Miden rollup tx kernel
-pub struct MidenTxKernelLibrary(CompiledLibrary);
-
-impl AsRef<CompiledLibrary> for MidenTxKernelLibrary {
-    fn as_ref(&self) -> &CompiledLibrary {
-        &self.0
-    }
-}
-
-impl From<MidenTxKernelLibrary> for CompiledLibrary {
-    fn from(lib: MidenTxKernelLibrary) -> Self {
-        lib.0
-    }
-}
-
-impl Default for MidenTxKernelLibrary {
-    fn default() -> Self {
-        let bytes = include_bytes!(concat!(env!("OUT_DIR"), "/assets/tx.masl"));
-        let contents = CompiledLibrary::read_from_bytes(bytes).expect("failed to read std masl!");
-        Self(contents)
-    }
-}
+#[cfg(feature = "bindings")]
+pub mod bindings;
+#[cfg(feature = "masl-lib")]
+pub mod masl;
