+++ sdk/sdk/src/lib.rs
@@ -1,4 +1,4 @@
 #![no_std]
 
+pub use miden_base_sys::bindings::tx::*;
 pub use miden_stdlib_sys::*;
-pub use miden_tx_kernel_sys::*;
