--- sdk/tx-kernel-sys/CHANGELOG.md
@@ -1,15 +1,0 @@
-# Changelog
-All notable changes to this project will be documented in this file.
-
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
-## [Unreleased]
-
-## [0.0.1](https://github.com/0xPolygonMiden/compiler/compare/miden-tx-kernel-sys-v0.0.0...miden-tx-kernel-sys-v0.0.1) - 2024-07-18
-
-### Other
-- set crates versions to 0.0.0, and `publish = false` for tests
-- rename `miden-sdk-tx-kernel` to `miden-tx-kernel-sys`
-- rename `miden-prelude` to `miden-stdlib-sys` in SDK
-- start guides for developing in rust in the book,
