--- sdk/tx-kernel-sys/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "miden-tx-kernel-sys"
-description = "Miden SDK transaction kernel"
-version = "0.0.1"
-rust-version.workspace = true
-authors.workspace = true
-repository.workspace = true
-categories.workspace = true
-keywords.workspace = true
-license.workspace = true
-readme.workspace = true
-edition.workspace = true
-
-[lib]
-crate-type = ["rlib"]
-
-[dependencies]
-miden-stdlib-sys = { path = "../stdlib-sys" }
