+++ tests/integration/expected/rust_sdk_account_test/miden_sdk_account_test.hir
@@ -165,7 +165,7 @@
         (func (export #get_wallet_magic_number) (result felt)
             (block 0
                 (let (v1 felt) (const.felt 0))
-                (let (v2 felt) (call #miden_tx_kernel_sys::get_id))
+                (let (v2 felt) (call #miden_base_sys::bindings::tx::get_id))
                 (let (v3 i64) (const.i64 42))
                 (let (v4 felt) (cast v3))
                 (let (v5 felt) (add.unchecked v4 v2))
@@ -217,7 +217,7 @@
                 (store v29 v8)
                 (let (v30 i32) (const.i32 16))
                 (let (v31 i32) (add.wrapping v5 v30))
-                (call #miden_tx_kernel_sys::add_asset v31 v5)
+                (call #miden_base_sys::bindings::tx::add_asset v31 v5)
                 (let (v32 u32) (bitcast v5))
                 (let (v33 u32) (add.checked v32 16))
                 (let (v34 u32) (mod.unchecked v33 4))
@@ -339,7 +339,7 @@
                 (store v7 v6)
                 (let (v8 i64) (const.i64 0))
                 (let (v9 felt) (cast v8))
-                (call #miden_tx_kernel_sys::get_inputs v6)
+                (call #miden_base_sys::bindings::tx::get_inputs v6)
                 (let (v10 u32) (bitcast v6))
                 (let (v11 u32) (mod.unchecked v10 4))
                 (assertz 250 v11)
@@ -664,7 +664,7 @@
                 (let (v6 i32) (sub.wrapping v4 v5))
                 (let (v7 (ptr i32)) (global.symbol #__stack_pointer))
                 (store v7 v6)
-                (call #miden_tx_kernel_sys::remove_asset v6 v0)
+                (call #miden_base_sys::bindings::tx::remove_asset v6 v0)
                 (let (v8 u32) (bitcast v6))
                 (let (v9 u32) (mod.unchecked v8 4))
                 (assertz 250 v9)
@@ -687,7 +687,7 @@
                     (param v1 felt)
                     (param v2 felt)
                     (param v3 i32)
-                (let (v5 felt) (call #miden_tx_kernel_sys::create_note v0 v1 v2 v3))
+                (let (v5 felt) (call #miden_base_sys::bindings::tx::create_note v0 v1 v2 v3))
                 (br (block 1 v5)))
 
             (block 1 (param v4 felt)
@@ -1601,7 +1601,7 @@
                 (br (block 12 v85 v87 v111 v101)))
         )
 
-        (func (export #miden_tx_kernel_sys::get_id)
+        (func (export #miden_base_sys::bindings::tx::get_id)
               (result felt)
             (block 0
                 (let (v1 felt) (call (#miden::account #get_id)))
@@ -1611,7 +1611,7 @@
                 (ret v0))
         )
 
-        (func (export #miden_tx_kernel_sys::get_inputs)
+        (func (export #miden_base_sys::bindings::tx::get_inputs)
               (param i32)
             (block 0 (param v0 i32)
                 (let (v1 i32) (const.i32 0))
@@ -1688,7 +1688,7 @@
                 (unreachable))
         )
 
-        (func (export #miden_tx_kernel_sys::add_asset)
+        (func (export #miden_base_sys::bindings::tx::add_asset)
               (param i32) (param i32)
             (block 0 (param v0 i32) (param v1 i32)
                 (let (v2 u32) (bitcast v1))
@@ -1733,7 +1733,7 @@
                 (ret))
         )
 
-        (func (export #miden_tx_kernel_sys::remove_asset)
+        (func (export #miden_base_sys::bindings::tx::remove_asset)
               (param i32) (param i32)
             (block 0 (param v0 i32) (param v1 i32)
                 (let (v2 u32) (bitcast v1))
@@ -1778,7 +1778,7 @@
                 (ret))
         )
 
-        (func (export #miden_tx_kernel_sys::create_note)
+        (func (export #miden_base_sys::bindings::tx::create_note)
               (param i32) (param felt) (param felt) (param i32) (result felt)
             (block 0
                     (param v0 i32)
