+++ tests/integration/expected/rust_sdk_account_test/miden_sdk_account_test.wat
@@ -48,11 +48,11 @@
   (import "std::crypto::hashes::blake3" "hash_1to1<0x0000000000000000000000000000000000000000000000000000000000000000>" (func $miden_stdlib_sys::stdlib::crypto::hashes::extern_blake3_hash_1to1 (;19;) (type 8)))
   (import "std::crypto::hashes::blake3" "hash_2to1<0x0000000000000000000000000000000000000000000000000000000000000000>" (func $miden_stdlib_sys::stdlib::crypto::hashes::extern_blake3_hash_2to1 (;20;) (type 9)))
   (import "std::crypto::dsa::rpo_falcon512" "rpo_falcon512_verify<0x0000000000000000000000000000000000000000000000000000000000000000>" (func $miden_stdlib_sys::stdlib::crypto::dsa::extern_rpo_falcon512_verify (;21;) (type 10)))
-  (import "miden::account" "get_id<0x0000000000000000000000000000000000000000000000000000000000000000>" (func $miden_tx_kernel_sys::externs::extern_account_get_id (;22;) (type 11)))
-  (import "miden::note" "get_inputs<0x0000000000000000000000000000000000000000000000000000000000000000>" (func $miden_tx_kernel_sys::externs::extern_note_get_inputs (;23;) (type 12)))
-  (import "miden::account" "add_asset<0x0000000000000000000000000000000000000000000000000000000000000000>" (func $miden_tx_kernel_sys::externs::extern_account_add_asset (;24;) (type 13)))
-  (import "miden::account" "remove_asset<0x0000000000000000000000000000000000000000000000000000000000000000>" (func $miden_tx_kernel_sys::externs::extern_account_remove_asset (;25;) (type 13)))
-  (import "miden::tx" "create_note<0x0000000000000000000000000000000000000000000000000000000000000000>" (func $miden_tx_kernel_sys::externs::extern_tx_create_note (;26;) (type 14)))
+  (import "miden::account" "get_id<0x0000000000000000000000000000000000000000000000000000000000000000>" (func $miden_base_sys::bindings::tx::externs::extern_account_get_id (;22;) (type 11)))
+  (import "miden::note" "get_inputs<0x0000000000000000000000000000000000000000000000000000000000000000>" (func $miden_base_sys::bindings::tx::externs::extern_note_get_inputs (;23;) (type 12)))
+  (import "miden::account" "add_asset<0x0000000000000000000000000000000000000000000000000000000000000000>" (func $miden_base_sys::bindings::tx::externs::extern_account_add_asset (;24;) (type 13)))
+  (import "miden::account" "remove_asset<0x0000000000000000000000000000000000000000000000000000000000000000>" (func $miden_base_sys::bindings::tx::externs::extern_account_remove_asset (;25;) (type 13)))
+  (import "miden::tx" "create_note<0x0000000000000000000000000000000000000000000000000000000000000000>" (func $miden_base_sys::bindings::tx::externs::extern_tx_create_note (;26;) (type 14)))
   (import "std::mem" "pipe_words_to_memory<0x0000000000000000000000000000000000000000000000000000000000000000>" (func $miden_stdlib_sys::stdlib::mem::extern_pipe_words_to_memory (;27;) (type 15)))
   (import "std::mem" "pipe_double_words_to_memory<0x0000000000000000000000000000000000000000000000000000000000000000>" (func $miden_stdlib_sys::stdlib::mem::extern_pipe_double_words_to_memory (;28;) (type 16)))
   (func $<<alloc::vec::into_iter::IntoIter<T,A> as core::ops::drop::Drop>::drop::DropGuard<T,A> as core::ops::drop::Drop>::drop (;29;) (type 17) (param i32)
@@ -142,7 +142,7 @@
   )
   (func $get_wallet_magic_number (;34;) (type 11) (result f32)
     (local f32)
-    call $miden_tx_kernel_sys::get_id
+    call $miden_base_sys::bindings::tx::get_id
     local.set 0
     i64.const 42
     call $miden_stdlib_sys::intrinsics::felt::extern_from_u64_unchecked
@@ -182,7 +182,7 @@
     i32.const 16
     i32.add
     local.get 0
-    call $miden_tx_kernel_sys::add_asset
+    call $miden_base_sys::bindings::tx::add_asset
     local.get 0
     f32.load offset=16
     local.set 1
@@ -292,7 +292,7 @@
     call $miden_stdlib_sys::intrinsics::felt::extern_from_u64_unchecked
     local.set 1
     local.get 0
-    call $miden_tx_kernel_sys::get_inputs
+    call $miden_base_sys::bindings::tx::get_inputs
     local.get 0
     i32.load
     local.set 2
@@ -440,7 +440,7 @@
     global.set $__stack_pointer
     local.get 1
     local.get 0
-    call $miden_tx_kernel_sys::remove_asset
+    call $miden_base_sys::bindings::tx::remove_asset
     local.get 1
     f32.load
     local.set 2
@@ -455,7 +455,7 @@
     local.get 1
     local.get 2
     local.get 3
-    call $miden_tx_kernel_sys::create_note
+    call $miden_base_sys::bindings::tx::create_note
   )
   (func $__rust_alloc (;45;) (type 24) (param i32 i32) (result i32)
     i32.const 1048576
@@ -1022,10 +1022,10 @@
       i32.store
     end
   )
-  (func $miden_tx_kernel_sys::get_id (;52;) (type 11) (result f32)
-    call $miden_tx_kernel_sys::externs::extern_account_get_id
+  (func $miden_base_sys::bindings::tx::get_id (;52;) (type 11) (result f32)
+    call $miden_base_sys::bindings::tx::externs::extern_account_get_id
   )
-  (func $miden_tx_kernel_sys::get_inputs (;53;) (type 17) (param i32)
+  (func $miden_base_sys::bindings::tx::get_inputs (;53;) (type 17) (param i32)
     (local i32 i32 i32)
     global.get $__stack_pointer
     i32.const 16
@@ -1055,7 +1055,7 @@
     local.get 1
     i32.load offset=12
     local.tee 3
-    call $miden_tx_kernel_sys::externs::extern_note_get_inputs
+    call $miden_base_sys::bindings::tx::externs::extern_note_get_inputs
     drop
     local.get 0
     i32.const 0
@@ -1071,7 +1071,7 @@
     i32.add
     global.set $__stack_pointer
   )
-  (func $miden_tx_kernel_sys::add_asset (;54;) (type 20) (param i32 i32)
+  (func $miden_base_sys::bindings::tx::add_asset (;54;) (type 20) (param i32 i32)
     local.get 1
     f32.load
     local.get 1
@@ -1081,9 +1081,9 @@
     local.get 1
     f32.load offset=12
     local.get 0
-    call $miden_tx_kernel_sys::externs::extern_account_add_asset
+    call $miden_base_sys::bindings::tx::externs::extern_account_add_asset
   )
-  (func $miden_tx_kernel_sys::remove_asset (;55;) (type 20) (param i32 i32)
+  (func $miden_base_sys::bindings::tx::remove_asset (;55;) (type 20) (param i32 i32)
     local.get 1
     f32.load
     local.get 1
@@ -1093,9 +1093,9 @@
     local.get 1
     f32.load offset=12
     local.get 0
-    call $miden_tx_kernel_sys::externs::extern_account_remove_asset
+    call $miden_base_sys::bindings::tx::externs::extern_account_remove_asset
   )
-  (func $miden_tx_kernel_sys::create_note (;56;) (type 23) (param i32 f32 f32 i32) (result f32)
+  (func $miden_base_sys::bindings::tx::create_note (;56;) (type 23) (param i32 f32 f32 i32) (result f32)
     local.get 0
     f32.load
     local.get 0
@@ -1114,7 +1114,7 @@
     f32.load offset=8
     local.get 3
     f32.load offset=12
-    call $miden_tx_kernel_sys::externs::extern_tx_create_note
+    call $miden_base_sys::bindings::tx::externs::extern_tx_create_note
   )
   (func $alloc::vec::Vec<T>::with_capacity (;57;) (type 20) (param i32 i32)
     (local i32 i32)
