+++ tests/integration/expected/rust_sdk_basic_wallet/rust_sdk_basic_wallet.hir
@@ -19,7 +19,7 @@
                 (let (v4 i32) (sub.wrapping v2 v3))
                 (let (v5 (ptr i32)) (global.symbol #__stack_pointer))
                 (store v5 v4)
-                (call #miden_tx_kernel_sys::add_asset v4 v0)
+                (call #miden_base_sys::bindings::tx::add_asset v4 v0)
                 (let (v6 i32) (const.i32 16))
                 (let (v7 i32) (add.wrapping v4 v6))
                 (let (v8 (ptr i32)) (global.symbol #__stack_pointer))
@@ -43,8 +43,8 @@
                 (let (v7 i32) (sub.wrapping v5 v6))
                 (let (v8 (ptr i32)) (global.symbol #__stack_pointer))
                 (store v8 v7)
-                (call #miden_tx_kernel_sys::remove_asset v7 v0)
-                (let (v9 felt) (call #miden_tx_kernel_sys::create_note v7 v1 v2 v3))
+                (call #miden_base_sys::bindings::tx::remove_asset v7 v0)
+                (let (v9 felt) (call #miden_base_sys::bindings::tx::create_note v7 v1 v2 v3))
                 (let (v10 i32) (const.i32 16))
                 (let (v11 i32) (add.wrapping v7 v10))
                 (let (v12 (ptr i32)) (global.symbol #__stack_pointer))
@@ -55,7 +55,7 @@
                 (ret))
         )
 
-        (func (export #miden_tx_kernel_sys::add_asset)
+        (func (export #miden_base_sys::bindings::tx::add_asset)
               (param i32) (param i32)
             (block 0 (param v0 i32) (param v1 i32)
                 (let (v2 u32) (bitcast v1))
@@ -100,7 +100,7 @@
                 (ret))
         )
 
-        (func (export #miden_tx_kernel_sys::remove_asset)
+        (func (export #miden_base_sys::bindings::tx::remove_asset)
               (param i32) (param i32)
             (block 0 (param v0 i32) (param v1 i32)
                 (let (v2 u32) (bitcast v1))
@@ -145,7 +145,7 @@
                 (ret))
         )
 
-        (func (export #miden_tx_kernel_sys::create_note)
+        (func (export #miden_base_sys::bindings::tx::create_note)
               (param i32) (param felt) (param felt) (param i32) (result felt)
             (block 0
                     (param v0 i32)
